Move PromptFallbacks into the Support namespace
